LPS-136283 Move to the correct folder
